refactor: separate PomoTimer widget
